Launcher2: Improve launcher app launching performance
To improve performance change the default behavior
of the launcher so that after onclick it immediately
launches the app instead of waiting for the icon animation
to finish. Also to improve performance simplify the icon
animation and start the animation on pointer down.

To change the default behavior back to its original
behavior set the POST_CLICK_ANIMATE
compile-time option to true.

Change-Id: Id42a1839ecb996f2ecc27e383ee2802219345ad3